crawler/pyproject.toml
[project]
name = "crawler"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "httpx>=0.28.1",
    "lxml>=6.0.2",
    "markdownify>=1.2.2",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pyyaml>=6.0.3",
    "readability-lxml>=0.8.4.1",
    "respx>=0.22.0",
    "tenacity>=9.1.4",
]

[dependency-groups]
dev = [
]
